Module feather_m0::atsamd21g18a::eic::config

Configuration n

Structs

FILTEN0R
Value of the field

FILTEN1R
Value of the field

FILTEN2R
Value of the field

FILTEN3R
Value of the field

FILTEN4R
Value of the field

FILTEN5R
Value of the field

FILTEN6R
Value of the field

FILTEN7R
Value of the field

R
Value read from the register

W
Value to write to the register

_FILTEN0W
Proxy

_FILTEN1W
Proxy

_FILTEN2W
Proxy

_FILTEN3W
Proxy

_FILTEN4W
Proxy

_FILTEN5W
Proxy

_FILTEN6W
Proxy

_FILTEN7W
Proxy

_SENSE0W
Proxy

_SENSE1W
Proxy

_SENSE2W
Proxy

_SENSE3W
Proxy

_SENSE4W
Proxy

_SENSE5W
Proxy

_SENSE6W
Proxy

_SENSE7W
Proxy

Enums

SENSE0R
Possible values of the field SENSE0

SENSE0W
Values that can be written to the field SENSE0

SENSE1R
Possible values of the field SENSE1

SENSE1W
Values that can be written to the field SENSE1

SENSE2R
Possible values of the field SENSE2

SENSE2W
Values that can be written to the field SENSE2

SENSE3R
Possible values of the field SENSE3

SENSE3W
Values that can be written to the field SENSE3

SENSE4R
Possible values of the field SENSE4

SENSE4W
Values that can be written to the field SENSE4

SENSE5R
Possible values of the field SENSE5

SENSE5W
Values that can be written to the field SENSE5

SENSE6R
Possible values of the field SENSE6

SENSE6W
Values that can be written to the field SENSE6

SENSE7R
Possible values of the field SENSE7

SENSE7W
Values that can be written to the field SENSE7